Description
Implementing the backpropagation learning algorithm using L2 regularization for a network with one hidden layer and two hidden neurons.  The size of the input layer is also set to two. Initializing the weights by drawing from a Gaussian distribution centered around zero, and learning the weights for the XOR data:

x1 = ( 1,-1,1,-1)

x2 = ( 1,1,-1,-1)

y = ( 1,0,0,1)